Break payment dialog widgets
