+++ lib/routes/home/widgets/payments_list/dialog/payment_details_dialog_lnurl_comment.dart
@@ -1,0 +1,35 @@
+import 'package:auto_size_text/auto_size_text.dart';
+import 'package:breez/bloc/account/account_model.dart';
+import 'package:breez/routes/home/widgets/payments_list/dialog/shareable_payment_row.dart';
+import 'package:breez_translations/breez_translations_locales.dart';
+import 'package:flutter/material.dart';
+
+class PaymentDetailsDialogLnurlComment extends StatelessWidget {
+  final PaymentInfo paymentInfo;
+  final AutoSizeGroup labelAutoSizeGroup;
+  final AutoSizeGroup valueAutoSizeGroup;
+
+  const PaymentDetailsDialogLnurlComment({
+    Key key,
+    this.paymentInfo,
+    this.labelAutoSizeGroup,
+    this.valueAutoSizeGroup,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final texts = context.texts();
+    final comment = paymentInfo.lnurlPayInfo?.comment?.trim();
+
+    if (comment == null || comment.isEmpty) {
+      return Container();
+    }
+
+    return ShareablePaymentRow(
+      title: texts.payment_details_dialog_share_comment,
+      sharedValue: comment,
+      labelAutoSizeGroup: labelAutoSizeGroup,
+      valueAutoSizeGroup: valueAutoSizeGroup,
+    );
+  }
+}
