+++ lib/routes/transactions/pos_payment_item.dart
@@ -7,7 +7,7 @@ import 'package:breez/routes/charge/currency_wrapper.dart';
 import 'package:breez/routes/charge/sale_view.dart';
 import 'package:breez/theme_data.dart' as theme;
 import 'package:breez/utils/date.dart';
-import 'package:breez/widgets/payment_details_dialog.dart';
+import 'package:breez/routes/home/widgets/payments_list/payment_details_dialog.dart';
 import 'package:breez/widgets/route.dart';
 import 'package:flutter/material.dart';
 import 'package:breez_translations/breez_translations_locales.dart';
