+++ lib/routes/withdraw_funds/tx_widget_with_info_message.dart
@@ -1,8 +1,7 @@
 import 'package:breez/bloc/reverse_swap/reverse_swap_model.dart';
-import 'package:breez/widgets/loader.dart';
-import 'package:breez/widgets/payment_details_dialog.dart';
-import 'package:flutter/material.dart';
+import 'package:breez/routes/home/widgets/payments_list/dialog/tx_widget.dart';
 import 'package:breez_translations/breez_translations_locales.dart';
+import 'package:flutter/material.dart';
 
 class TxWidgetWithInfoMsg extends StatelessWidget {
   const TxWidgetWithInfoMsg({
